Feat: 닉네임 검사 공통컴포넌트 분리 및 쿼리스트링으로 로그인방식 따라 분기
