+++ src/components/checkNickname/styles.ts
@@ -1,0 +1,5 @@
+import styled from "@emotion/styled";
+
+export const NoticeContainer = styled.div`
+  margin: 8px 0 18px;
+`;
